+++ framework/model/src/main/java/is/codion/framework/model/package-info.java
@@ -23,6 +23,7 @@
  * <li>{@link is.codion.framework.model.EntityApplicationModel#SAVE_DEFAULT_USERNAME}
  * <li>{@link is.codion.framework.model.EntityEditModel#POST_EDIT_EVENTS}
  * <li>{@link is.codion.framework.model.EntityEditModel.EntityEditor#PERSIST_FOREIGN_KEYS}
+ * <li>{@link is.codion.framework.model.EntityQueryModel#LIMIT}
  * <li>{@link is.codion.framework.model.EntityTableModel#ON_INSERT}
  * <li>{@link is.codion.framework.model.EntityTableModel#HANDLE_EDIT_EVENTS}
  * <li>{@link is.codion.framework.model.EntityTableModel#ORDER_QUERY_BY_SORT_ORDER}