+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - AbstractEntityTableModel.orderByFromSortModel() added.
 - EntityEditModel.EntityEditor.ValueEditor renamed EditorValue.
 - DefaultEntityApplicationModel, entityModels now initialized via constructor parameter, related refactoring.
+- EntityQueryModel.LIMIT configuration value added.
 ### is.codion.swing.framework.ui
 - EntityTablePanel.orderQueryBySortOrder() moved to model.
 - EntityPanel.Builder, major simplification and related refactoring.